personalization-model/requirements.txt
flask==3.0.3
boto3==1.34.69
sagemaker==2.224.0
scikit-learn==1.4.2
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.2.0
